Move game path browser into separate file
