+++ src/CMakeLists.txt
@@ -262,6 +262,8 @@ set(core_sources
     ui/episode_end_sequence.hpp
     ui/fps_display.cpp
     ui/fps_display.hpp
+    ui/game_path_browser.cpp
+    ui/game_path_browser.hpp
     ui/high_score_list.cpp
     ui/high_score_list.hpp
     ui/hud_renderer.cpp
